Resolved issue #1:
Hopefully this will prevent the creatures from re-spawning at the location of their death,
and force them to spawn at their spawn home point, how it was intended in the first place.

Also fixed a small mistake in the 'Creature.h', and created a new message to be displayed
when the player tries to re-spawn a creature that is not dead (there was no error message before).

This issue was resolved and published by LordUsagi, most credit should go to him.