Resolved issue #357 - Generic spells
Someone reported that the item called "Living Ruby Pendant" casts an aura that restores
health depending on ones healing bonus or spell power. Clearly this is not how this item
is designed to work, so after only a guess and a few minutes of testing it became obvious
that this problem is not only limited to this particular item. Other spells/auras of the
spell family 'generic' are having their spell effects scaled with the players healing bonus.

This issue was resolved by LordUsagi. Much needed thanks goes to him for making a lot of spells
and items like wands balanced and correct.