Resolved a lot of issues with hunters pets:
The first issue that was resolved involved situations where a pet was being revived
on the location of it's death, possibly 'aggroing' the mobs still around it.
Also, after it re-spawned it did not follow it's owner.

Pets are now properly revived next to their owners and should follow them.
They also no longer need to be called or summoned by the player in order to revive them.

Following up upon this we discovered a couple of uncomfortable things about pets:

- They could have been dismissed while being dead if their corpse was in the world
- You could have called your pet to your location when it was dead and de-spawned.
- The pet could have been revived after being dismissed and it would spawn with minimum
  health just as if it was dead and then revived.

Fixes a possible crash involved with pets and food items. In the case that a selected item
wasn't found, a server crash would occur. This helps to improve stability and server up time.

There was also a small visual bug that happened when a pet would last hit a target.
It would sometimes mess up the action bar and cause inconvenience, such as remaining
stationary at the victims corpse.

Unrelated to fixes and bugs, I've renamed some spell effects in the 'SharedDefines' and
'SpellEffects' files as they have not been named with their actual names.

This commit was written and put together by LordUsagi and myself.
Although we have tested this in many situations, being that there is more then a few things
changed and there is a new system for tracking some pet info please feel free to test more yourself
and report any unusual behaviour or bugs to the issue tracker.