Resolved issue #481 - Stealth & Vanish:
After commit 4e4f88f, it was reported that stealth and vanish were a bit harder to cast in
some situations. It would take a few clicks on the spell icon for the spell to cast, because of
the interruption notifications sent by the server.

The shadowmeld bug (issue #261) is now resolved a little different.
When initiating combat and shadowmelding right after that, the stealth will be allowed to
persist until the next spell is being cast. This now works for auto-repeat spells such as
hunters regular ranged attacks.

This has been resolved only for auto-repeat spells, but I think normal casting works fine.
Might be a bit bloated and not fully optimized being placed in a high traffic function,
but this will be fine until more serious code restructuring can be done.