Resolved issue #159 - Food & Drinks:
Food regeneration effects were permitted to stack, and this did not happen every time,
which reveals more inconsistencies in the core. Interesting enough this issue did not affect drinks
that regenerate a fixed amount of power. Note that I did not test the ones that regenerate percentages.

Okay, so the complete fix for this issue is not brought by this commit alone.
One of the latest commits (Player stats) restructured some code to disable
the core from adding/stacking auras and applying their modifiers if the
aura effects had the same ID. This prevented drinks and food that had the same effect from stacking.

As can be seen in the change-log, I've just added a check for what I believe are types of
auras that heal by percentage (SPELL_AURA_OBS_MOD_HEALTH). Because this was missing it caused
items such as Graccu's Mince Meat Fruitcake and Festival Dumplings's effects to stack together.