Resolved issue #110 - Pets & Spellcasting:
The previous code created a bug where pets would remain standing on the location
they were casting, instead of returning to their owner. This was happening because
the pets would loose their 'Follow' state, and would never have it reapplied.
In fact, this seems to be the wrong state to remove to begin with.

Written by LordUsagi.
Edited and published by Yooki.