Fixed some quests, and added padding to event_scripts for some.
In this commit I'm fixing quests, which have Quest Item, and the
quest item casts a spell and that spell has effect SendEvent.

Did I confuse you ? :)

Also adding padding to event_scripts, if the quest is working, this
is very useful for filtering when retrieving unprocessed quests from sql.

Fixed Quests that were NOT working:

- 11147: Unleash the Raptors
- 9931:  Returning the Favor
- 10256: Finding the Keymaster
- 9444:  Defile Uther's Tomb
- 10182; Battle-Mage Dathric
- 10306; Conjurer Luminrath
- 10307; Cohlien Frostweaver