Health should show up in percents not in values (configurable, read description)
Health on 2.4.3 should be in %, there are exceptions:
- you see your own health and your pets health
- as well you see your party/raid members' & their pets' health

Otherwise, you're supposed to see all health of units in percents.
This option is blizzlike and is turned on by default.

It's been added as "HealthInPercents" to the configuration file oregoncore.conf.
You should update it or you can append the "HealthInPercents" option to your
existing one. Set it to 0 to disable this feature.

Note: you can do this live, just save the configuration file and run .reload config.

References:

- http://wowwiki.wikia.com/Patch_3.0.2_%28undocumented_changes%29
- https://bitbucket.org/smolderforge/core/issue/7/server-is-giving-the-client-player-mob
- Any video from youtube recorded at official servers

----

A bug was fixed by this commit as well: Updates on pets will now be properly send to group members
in case of when player was in group and tamed/abandoned a pet later.